string methods

+++ arithmetics.py
@@ -4,7 +4,7 @@ result  = (x + y) * 20
 print(result)
 
 num = input('Number: ')
-print(int(num) * 5)
+print(float(num) * 5)
 #The int converts str to a number. 
 #The 'int' can be changed to float which will return the number with a decimal.
